Split the model tests into individual files
